backend/pyproject.toml
[tool.poetry]
name = "sb"
version = "0.1.0"
description = ""
authors = ["Klukvas <pavlenkoandrey56@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
